♻️ JNXQR1 task: split overlay project helpers
